// packages/voice/lib/enums/HttpMethod.ts
/**
 * Enum representing HTTP methods used for making web requests.
 */
export enum HttpMethod {
  /**
   * The HTTP GET method, used for retrieving data from a server.
   */
  GET = 'GET',

  /**
   * The HTTP POST method, used for sending data to a server to create or update resources.
   */
  POST = 'POST',
}